okay, it should actually be fixed now

+++ src/requirements.cpp
@@ -467,7 +467,7 @@ std::vector<std::string> requirement_data::get_folded_list( int width,
                 if( item::count_by_charges( item_id ) ) {
                     qty = crafting_inv.charges_of( item_id, std::numeric_limits<long>::max(), filter );
                 } else {
-                    qty = crafting_inv.amount_of( item_id, false, std::numeric_limits<long>::max(), filter );
+                    qty = crafting_inv.amount_of( item_id, false, std::numeric_limits<int>::max(), filter );
                 }
                 text = item::count_by_charges( item_id ) ?
                        string_format( _( "%s (%d of %ld)" ), item::nname( item_id ), component.count * batch, qty ) :
